Fix Xinerama detection at runtime.
The problem was in the logic, namely the comparison
which got carried from the module reorganization.

The change is done in such a way that instead of

if (Xinerama_works) then { do_it(); }

now it is

if (!Xinerama_works) then { fallback; return; }
if (!Xinerama_query) then { fallback; return; }
/* We know Xinerama worked. */

and so, it is primarily just movement of code (implies
indentation) and fixing of conditionals.