Make sure KeyReleased keys are actually not pressed (Bug #432887).
Some KeyRelease events were generated by X's AutoRepeat, even if
the key wasn't actually released. Using XQueryKeymap we check the
state of the key. If it isn't a true KeyRelease the event is removed
from the stack, else it executes the action associated.

This solves bug described in
https://bugs.launchpad.net/superkb/+bug/432887/
by Octavio Alvarez as:

> When the bound key is held down before the Super key, Superkb
> will perform the associated action multiple times.
>
> Apparently, this is because autorepeat is not getting disabled
> for the pressed down key.