Fix wrong display of multibyte characters.
... like &ntilde; on latin american keyboards.

Multiple-byte characters were counted as multiple
characters by strlen() and thus, painted as such.

The fix consists on making an mbstrlen() function
based on mblen() and using it. However, mblen()
depends on the locale, so we added a setlocale()
call to load it from the environment variables.

The user must load Superkb with something like
$ LC_ALL=en_US.UTF-8 superkb
or
$ LC_CTYPE=en_US.UTF-8 superkb