Reworked XNextEventWithTimeout.
New way of signaling autorepeats by the X11 servers broke Superkb.
Before we had a lot of KeyPresses and a single KeyRelease. Now we
have a lot of KeyPress/KeyRelease combinations. We must peek into
the next X event to see if it matches a KeyPress. If so, we must
ignore both.

We had to invert the logic for QLength() == 0 management and enclose
everything in a loop to continue if the events were swallowed.

This is to only emulate the old way before the rest of the codebase.

See: https://stackoverflow.com/questions/2100654

Signed-off-by: Octavio Alvarez <alvarezp@alvarezp.ods.org>