Change all -I CFLAGS to -isystem
We are not expert C programmers so we use compile using the GCC
--pedantic-errors option. This flags helps us find some errors
related to our C-related ignorance before they get pushed publicly.

In Superkb issue #7 [1] on Github, it came up to our attention that
Superkb fails to build using the still-unreleased GCC 6. A Debian
test found this and a bug against the Superkb bug was reported [2].

It turned out the failure to build lies inside GLib code. A bug was
reported to the GLib team [3] and a possible workaround provided but
they considered this not to be a bug and refused to fix it.

So, instead of root-fixing the issue we are forced to write a work
around. We will silence all pedantic errors for system libraries. GCC
does not show warnings or errors for system libraries. For that to
apply, included files must be loaded using the -isytem option instead
of the typical -I.

In this patch we replace all -I to -isystem in the pkg-config output
for the CFLAGS variable so GCC bypasses system-library mistakes.

[1] https://github.com/alvarezp/superkb/issues/7
[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=811685
[3] https://bugzilla.gnome.org/show_bug.cgi?id=761119