Moved the help text outside of main.c
In Debian bug #929005, Helmut Grohen correctly points out that
Superkb is not cross-buildable because of the use of help2man.

Help2man works by running the built binary with -h and -v to
find out the help text and version and build a manpage from that.
If we are cross-building Superkb we probably don't have access
to a host with the target architecture, so we don't have a way
for help2man to run the application binary to get needed texts.

This will be fixed in two steps. First (this one) will move
the help text outside of main.c into a .h in a specific format
that can be processed by a "help stub" written in Bash. The help
stub will be called by help2man instead of the application binary.
The help stub will be included in the second commit of this fix.

For now we are just getting the help text out of main.c and
nothing should break at all.