Add and use a help2man stub
To help Helmut's attempt at having Debian fully cross-compile
we need to prevent help2man from running the application binary
because if we are cross-building it means we don't have access
to a host with the target architecture, and thus, help2man
cannot run the application binary to get the help text and the
application version.

In a previous commit we moved the help text outside of main.c
so the help stub can read it and generate the help text for
help2man to use. This time we are actually introducing the
help stub and making help2man use it.