help2man/superkb: way cleaner method to use main-help-message.h
Using search and replace was an ugly hack. It forces the .h
file use a specific format that can easily be broken.

Instead, use cpp itself to parse the file. Without the echo
file we only get comments back from cpp -E so we just append
an "echo -e HELP_MESSAGE" line on the fly to the input to cpp
so it generates a valid Bash one-liner that echoes the parsed
help message.